test: split native package manifest coverage
