+++ tests/tooling/package-manifests.test.ts
@@ -588,36 +588,3 @@ test("workspace TypeScript package builds use vp pack", () => {
   const rootTasks = readRepoFile("tools/config/vite-plus/tasks/build.ts");
   assert.match(rootTasks, /vscodeExtensionPackageBin\("vite-plus", "vp"\)[\s\S]*pack/);
 });
-
-test("fresco-native publishes bundled binaries directly from the root package", () => {
-  const frescoNativePackage = JSON.parse(
-    fs.readFileSync(path.join(root, "npm/fresco-native/package.json"), "utf-8"),
-  ) as {
-    files?: string[];
-    scripts?: Record<string, string>;
-  };
-  const vizeNativePackage = JSON.parse(
-    fs.readFileSync(path.join(root, "npm/native/package.json"), "utf-8"),
-  ) as {
-    scripts?: Record<string, string>;
-  };
-
-  assert.deepEqual(frescoNativePackage.files, ["index.js", "index.d.ts", "*.node"]);
-  assert.equal(frescoNativePackage.scripts?.prepublishOnly, undefined);
-  assert.equal(
-    frescoNativePackage.scripts?.["build:ci"],
-    "napi build --platform --profile ci --manifest-path ../../crates/vize_fresco/Cargo.toml -p vize_fresco --features napi --output-dir .",
-  );
-  assert.equal(
-    vizeNativePackage.scripts?.["build:ci"],
-    "napi build --platform --profile ci --manifest-path ../../crates/vize_vitrine/Cargo.toml -p vize_vitrine --features napi,legacy --output-dir . && node ./scripts/sync-entrypoint.mjs",
-  );
-
-  const frescoNativeLoader = fs.readFileSync(
-    path.join(root, "npm/fresco-native/index.js"),
-    "utf-8",
-  );
-  assert.match(frescoNativeLoader, /spawnSync\("ldd", \["--version"\]/);
-  assert.doesNotMatch(frescoNativeLoader, /execSync\("which ldd"\)/);
-  assert.doesNotMatch(frescoNativeLoader, /readFileSync\(lddPath/);
-});
